+++ dynts/maths/__init__.py
@@ -1,11 +1,13 @@
 from itertools import izip
-from utils import *
 
 
 class BasicStatistics(object):
     
     def __init__(self, ts):
         self.ts = ts
+        
+    def count(self):
+        return self.ts.count()
         
     def calculate(self):
         tseries = self.ts
@@ -21,6 +23,70 @@ class BasicStatistics(object):
                 'max': list(tseries.max()),
                 }
         return data
+    
+    
+class pivottable(object):
+    
+    def __init__(self, data, default = 'latest'):
+        self.default = default
+        self.data = data
+        if data:
+            self.names = data['names']
+            self.defaultname = self.names[0]
+            d = self.default
+            self._names = dict(((name,{d:v}) for name,v in izip(data['names'],data[d])))
+        
+    def __get_fields(self):
+        return self.data.keys()
+    fields = property(__get_fields)
+        
+    def get(self, code, name = None):
+        dnames = self._names
+        
+        # First we check if code is a name
+        if not name:
+            v = dnames.get(code,None)
+            if v:
+                return v[self.default]
+            name = self.defaultname
+            
+        # Check if name is available, otherwise return None
+        nd = dnames.get(name,None)
+        if not nd:
+            return None
+        
+        v = nd.get(code,None)
+        if v:
+            return v
+        
+        v = self.data.get(code,None)
+        if v:
+            for nam,val in izip(self.names,v):
+                dnames[nam][code] = val
+        
+        v = nd.get(code,None)
+        if v:
+            return v
+        
+        func = getattr(self,'calculate_{0}'.format(code),None)
+        if func:
+            return func(name)
+    
+    def calculate_prange(self, name):
+        '''Latest value as percentage in range'''
+        lat  = self.get('latest',name)
+        min  = self.get('min',name)
+        max  = self.get('max',name)
+        dd   = max - min
+        if dd:
+            return 100.*(lat-min)/dd
+        else:
+            return 0.
+        
+    def calculate_range(self, name, significant = 4):
+        min  = self.get('min',name)
+        max  = self.get('max',name)
+        return [min,max]
     
 
 class SimpleStatisticsTable(object):