+++ dynts/test.py
@@ -1,10 +1,15 @@
 import unittest
-import numpy as np
 from datetime import date
 from itertools import izip
+import types
+
+import numpy as np
 
 from dynts import timeseries, evaluate, tsname
 from dynts.utils.populate import populate, datepopulate, randomts
+
+
+TextTestRunner = unittest.TextTestRunner
 
 
 class TestCase(unittest.TestCase):
@@ -26,6 +31,7 @@ class TestCase(unittest.TestCase):
         return dates,data
         
     def getts(self, returndata = False, delta = 1, cols = 1, size = 100):
+        '''Return a timeseries filled with random data'''
         dates,data = self.getdata(size,cols,delta)
         ts   = self.timeseries(name = 'test', date = dates, data = data, backend = self.backend)
         self.assertEqual(ts.shape,(size,cols))
@@ -82,3 +88,57 @@ class TestCase(unittest.TestCase):
                 self.assertEqual(a, b)
             else:
                 self.assertAlmostEqual(a,b)
+
+
+        
+
+class TestSuite(unittest.TestSuite):
+    pass
+
+
+class TestSuiteRunner(object):
+    '''A suite runner with twisted if available.'''
+    
+    def __init__(self, verbosity = 1):
+        self.verbosity = verbosity
+        
+    def setup_test_environment(self):
+        pass
+    
+    def teardown_test_environment(self):
+        pass
+    
+    def run_tests(self, modules):
+        self.setup_test_environment()
+        suite = self.build_suite(modules)
+        self.run_suite(suite)
+    
+    def close_tests(self, result):
+        self.teardown_test_environment()
+        return self.suite_result(suite, result)
+    
+    def build_suite(self, modules):
+        loader = TestLoader()
+        return loader.loadTestsFromModules(modules)
+        
+    def run_suite(self, suite):
+        return TextTestRunner(verbosity = self.verbosity).run(suite)
+    
+    def suite_result(self, suite, result, **kwargs):
+        return len(result.failures) + len(result.errors) 
+     
+
+class TestLoader(unittest.TestLoader):
+    suiteClass = TestSuite
+    
+    def loadTestsFromModules(self, modules):
+        """Return a suite of all tests cases contained in the given module"""
+        tests = []
+        for module in modules:
+            for name in dir(module):
+                obj = getattr(module, name)
+                if (isinstance(obj, (type, types.ClassType)) and
+                    issubclass(obj, unittest.TestCase)):
+                    tests.append(self.loadTestsFromTestCase(obj))
+        return self.suiteClass(tests)
+