+++ dynts/tests/regression/functions/tests.py
@@ -1,0 +1,13 @@
+from regression.functions import functions
+
+class TestMeanFunction(functions.SimpleFunctionTest):
+    function = 'ma'
+
+class TestMaxFunction(functions.SimpleFunctionTest):
+    function = 'max'
+    
+class TestMedFunction(functions.SimpleFunctionTest):
+    function = 'med'
+
+class TestMinFunction(functions.SimpleFunctionTest):
+    function = 'min'