+++ dynts/utils/__init__.py
@@ -4,3 +4,4 @@ from wrappers import *
 from skiplist import *
 from section import *
 from encoding import *
+from numbers import *